[FIX] split by widget
